- Added migration to Restaurant - has field createdById (which user created it). Only this user can for example delete this restaurant - added ClaimsPrincipal user to restaurant service, it's gonna be refactored soon - Added UserContextService - soon this class gonna handle user data in restaurant service

+++ Services/RestaurantService.cs
@@ -1,10 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Net.NetworkInformation;
+using System.Security.Claims;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Controllers;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Entitites;
@@ -17,9 +21,9 @@ namespace RestaurantAPI.Services
     {
         ActionResult<RestaurantDto> GetById(int id);
         IEnumerable<RestaurantDto> GetAllRestaurants();
-        int CreateRestaurant(CreateRestaurantDto dto);
-        void DeleteRestaurant(int id);
-        void UpdateRestaurant(int id, RestaurantModifyDto dto);
+        int CreateRestaurant(CreateRestaurantDto dto, int userId);
+        void DeleteRestaurant(int id, ClaimsPrincipal user);
+        void UpdateRestaurant(int id, RestaurantModifyDto dto, ClaimsPrincipal user);
     }
 
 
@@ -28,12 +32,14 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
         public ActionResult<RestaurantDto> GetById(int id)
@@ -65,9 +71,10 @@ namespace RestaurantAPI.Services
             return restaurantsDtos;
         }
 
-        public int CreateRestaurant(CreateRestaurantDto dto)
+        public int CreateRestaurant(CreateRestaurantDto dto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
 
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
@@ -75,7 +82,7 @@ namespace RestaurantAPI.Services
             return restaurant.Id;
         }
 
-        public void DeleteRestaurant(int id)
+        public void DeleteRestaurant(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Some error in delete restaurant");
 
@@ -86,18 +93,31 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = _authorizationService
+                .AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException();
+
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
         }
 
-        public void UpdateRestaurant(int id, RestaurantModifyDto dto)
+        public void UpdateRestaurant(int id, RestaurantModifyDto dto, ClaimsPrincipal user)
         {
+
             var restaurant = _dbContext
                 .Restaurants
                 .FirstOrDefault(r => r.Id == id);
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService
+                .AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException();
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
